Fully encapsulate the events module and declutter connector
