+++ servo/__init__.py
@@ -23,18 +23,20 @@ from .events import (
     Preposition,
     EventError,
     CancelEventError,
-)
-from .connector import (
-    Optimizer,
-    BaseConfiguration,
-    Connector,
-    metadata,
+    create_event,
     event,
     before_event,
     on_event,
     after_event,
     event_handler,
 )
+from .connector import (
+    Optimizer,
+    BaseConfiguration,
+    Connector,
+    metadata,
+)
 
-# Pull the types up to the top level
+# Pull the types & utilities up to the top level
 from .types import *
+from .utilities import *
