+++ servo/utilities/__init__.py
@@ -1,7 +1,6 @@
 from .strings import *
 from .duration_str import *
 from .pydantic import *
-from .progress import *
 from .signal_handler import *
 from .yaml import *
 from .key_paths import *
